Make ParseContext from JsonPath.using thread safe #187
