+++ json-path/src/main/java/com/jayway/jsonpath/JsonPath.java
@@ -16,7 +16,7 @@ package com.jayway.jsonpath;
 
 
 import com.jayway.jsonpath.internal.EvaluationContext;
-import com.jayway.jsonpath.internal.JsonContext;
+import com.jayway.jsonpath.internal.ParseContextImpl;
 import com.jayway.jsonpath.internal.Path;
 import com.jayway.jsonpath.internal.PathRef;
 import com.jayway.jsonpath.internal.Utils;
@@ -499,7 +499,7 @@ public class JsonPath {
      */
     @SuppressWarnings({"unchecked"})
     public static <T> T read(String json, String jsonPath, Predicate... filters) {
-        return new JsonContext().parse(json).read(jsonPath, filters);
+        return new ParseContextImpl().parse(json).read(jsonPath, filters);
     }
 
 
@@ -515,7 +515,7 @@ public class JsonPath {
     @SuppressWarnings({"unchecked"})
     @Deprecated
     public static <T> T read(URL jsonURL, String jsonPath, Predicate... filters) throws IOException {
-        return new JsonContext().parse(jsonURL).read(jsonPath, filters);
+        return new ParseContextImpl().parse(jsonURL).read(jsonPath, filters);
     }
 
     /**
@@ -529,7 +529,7 @@ public class JsonPath {
      */
     @SuppressWarnings({"unchecked"})
     public static <T> T read(File jsonFile, String jsonPath, Predicate... filters) throws IOException {
-        return new JsonContext().parse(jsonFile).read(jsonPath, filters);
+        return new ParseContextImpl().parse(jsonFile).read(jsonPath, filters);
     }
 
     /**
@@ -543,7 +543,7 @@ public class JsonPath {
      */
     @SuppressWarnings({"unchecked"})
     public static <T> T read(InputStream jsonInputStream, String jsonPath, Predicate... filters) throws IOException {
-        return new JsonContext().parse(jsonInputStream).read(jsonPath, filters);
+        return new ParseContextImpl().parse(jsonInputStream).read(jsonPath, filters);
     }
 
 
@@ -555,13 +555,15 @@ public class JsonPath {
 
 
     /**
-     * Creates a {@link ParseContext} that can be used to parse a given JSON input.
+     * Creates a {@link ParseContext} that can be used to parse JSON input. The parse context
+     * is as thread safe as the underlying {@link JsonProvider}. Note that not all JsonProvider are
+     * thread safe.
      *
      * @param configuration configuration to use when parsing JSON
      * @return a parsing context based on given configuration
      */
     public static ParseContext using(Configuration configuration) {
-        return new JsonContext(configuration);
+        return new ParseContextImpl(configuration);
     }
 
     /**
@@ -572,7 +574,7 @@ public class JsonPath {
      */
     @Deprecated
     public static ParseContext using(JsonProvider provider) {
-        return new JsonContext(Configuration.builder().jsonProvider(provider).build());
+        return new ParseContextImpl(Configuration.builder().jsonProvider(provider).build());
     }
 
     /**
@@ -583,7 +585,7 @@ public class JsonPath {
      * @return a read context
      */
     public static DocumentContext parse(Object json) {
-        return new JsonContext().parse(json);
+        return new ParseContextImpl().parse(json);
     }
 
     /**
@@ -594,7 +596,7 @@ public class JsonPath {
      * @return a read context
      */
     public static DocumentContext parse(String json) {
-        return new JsonContext().parse(json);
+        return new ParseContextImpl().parse(json);
     }
 
     /**
@@ -605,7 +607,7 @@ public class JsonPath {
      * @return a read context
      */
     public static DocumentContext parse(InputStream json) {
-        return new JsonContext().parse(json);
+        return new ParseContextImpl().parse(json);
     }
 
     /**
@@ -616,7 +618,7 @@ public class JsonPath {
      * @return a read context
      */
     public static DocumentContext parse(File json) throws IOException {
-        return new JsonContext().parse(json);
+        return new ParseContextImpl().parse(json);
     }
 
     /**
@@ -627,7 +629,7 @@ public class JsonPath {
      * @return a read context
      */
     public static DocumentContext parse(URL json) throws IOException {
-        return new JsonContext().parse(json);
+        return new ParseContextImpl().parse(json);
     }
 
     /**
@@ -638,7 +640,7 @@ public class JsonPath {
      * @return a read context
      */
     public static DocumentContext parse(Object json, Configuration configuration) {
-        return new JsonContext(configuration).parse(json);
+        return new ParseContextImpl(configuration).parse(json);
     }
 
     /**
@@ -649,7 +651,7 @@ public class JsonPath {
      * @return a read context
      */
     public static DocumentContext parse(String json, Configuration configuration) {
-        return new JsonContext(configuration).parse(json);
+        return new ParseContextImpl(configuration).parse(json);
     }
 
     /**
@@ -660,7 +662,7 @@ public class JsonPath {
      * @return a read context
      */
     public static DocumentContext parse(InputStream json, Configuration configuration) {
-        return new JsonContext(configuration).parse(json);
+        return new ParseContextImpl(configuration).parse(json);
     }
 
     /**
@@ -671,7 +673,7 @@ public class JsonPath {
      * @return a read context
      */
     public static DocumentContext parse(File json, Configuration configuration) throws IOException {
-        return new JsonContext(configuration).parse(json);
+        return new ParseContextImpl(configuration).parse(json);
     }
 
     /**
@@ -682,7 +684,7 @@ public class JsonPath {
      * @return a read context
      */
     public static DocumentContext parse(URL json, Configuration configuration) throws IOException {
-        return new JsonContext(configuration).parse(json);
+        return new ParseContextImpl(configuration).parse(json);
     }
 
     private <T> T resultByConfiguration(Object jsonObject, Configuration configuration, EvaluationContext evaluationContext) {
